% Auto-generated: do not edit by hand
\name{icicle}

\alias{icicle}

\title{Icicle component}

\description{

}

\usage{
icicle(id=NULL, label=NULL, value=NULL, data=NULL, low=NULL,
high=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}

\item{data}{Named list. }

\item{low}{Numeric. }

\item{high}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

